Add G38.2 gcode
grbl implements G38.2 for three axes; this only implements the z axis.
G38.2 is supposed to finish at the touch point without altering the
home position, the user would use G92 for that. We don’t have G92
(beyond the scope of this :) ) so this zeroes the z axis at the touch
point. If the touch point is not reached, the ‘STOP’ command is issued
to halt further execution without altering the zero. An example of
operation would be to execute this command file:

G20
G90
G38.2 Z-.15 F1
G20
G90
M02

The probe is assumed to be connected to AUX4, which is set at the top
of CNC_Functions.h . Ground is attached to the router (bit) and AUX4 to
the touch plate. AUX4 is configured as INPUT_PULLUP and is pulled LOW
when the probe touches the plate.