Put back in abs() on final Number Of Steps
Had to put back in the abs() on final number of steps. It appears that
it needs to be negitive some times to convey information about which way
the arc is being cut.

This should be handled in a more clear way because why the abs() is
needed is not clear from reading the code.