Parameters as references
Arduino pointer size is 8 bits, so any parameter larger than a byte or char, should be passed by reference. In cases where these are read-only parameters, they are qualified as const to prevent accidental usages. There are a few places they are modified locally, in which case they have been left alone so that a function-local copy is made.