abs usage issue in Axis::error
Due to the abs function being implemented as a DEFINE, the argument to it will be referenced 3 times. Axis:error was calling the encoder.read function within the call to abs. This would cause that function to be executed 3 times, potentially returning the wrong value if the value read each time was different. Modified to store the encoder.read valuel and then call abs with the stored value.