Constrain ZAxis Feedrate Based on Maximum RPM of Z-Axis Motor
The feed rate for all zaxis movements had been hardcoded at 45mm/min. Best I can tell, my motor was unable to keep up with this rate.  This caused jerky movements by the zaxis and resulted in it not completing its movement before the next command was executed.  This would cause plunge depths to be incomplete, and on retraction would often lead to dragging cuts when moving to the next location.

I calculated the maximum rpm of my zaxis by pushing the feed rate to absurd levels and then timing the rpms of the motor.

The max feedrate has to limited by the motor rpm, not a hardcoded rate, as other routers may have higher or lower zaxis pitches.