Fix Calculation of Angle in Arc Function
The calculation of theta was wrong.  A good example was theta would equal zero for any cut in quadrant 2 ending in quadrant 3, where the start and end angles were the same.

To fix:
1. flip the calculation so it is ending point minus starting point.
2. The values of atan2 are between -pi and pi.  So we need to normalize to 2pi

No longer need to specifically catch the edge case of a full circle as this is now caught as part of the regular processing.