RingBuffer updates and add numberOfLines
In RingBuffer::read(), move _incrementBeginning() inside else statement.  If the buffer is empty, we definitely don't want to move the beginning pointer.  Add RingBuffer::numberOfLines() to count the number of full lines in the buffer, update RingBuffer::readLine() to use this count instead of looking for a full line.  Update RingBuffer::print() to dump the buffer contents, update RingBuffer::_incrementEnd() to use RingBuffer::print() for buffer info and contents instead of its own code.  Fix comments for RingBuffer::spaceAvailable() and RingBuffer::length().