Command String trim improvements
A recent update added the use of string .trim() in interpretCommandString(), but the trimmed string was only used to determine whether the command was a blank line.  This uses extra memory by holding both the untrimmed and trimmed strings in memory, and the remaining functions in interpretCommandString() would benefit from being passed a trimmed string.  Moved the string trim() out to the main loop() function before passing to interpretCommandString().