Watchdog - never add EOL to buffer
Change _watchdog() to request new code from Ground Control when necessary, but never add an end-of-line character to the buffer - this is how the buffer gets filled with extra lines and eventually overflows.

Also change timeout to an unsigned long to make sure comparison with millis() and lastRan is successful.