Update signalReady to check space available
Added check in signalReady for space available in buffer before requesting another line of code from Ground Control.  Increased maximum number of lines in the buffer to 4.  (We want to limit the maximum number of lines in the buffer because Maslow will continue to run the remaining lines in the buffer after the user presses the "Halt" button in Ground Control.