Delete the broken "Test Electronics" firmware
During the beta testing process we had so many issues that it made sense to have a firmware branch which just made the motors go and printed the words "connection good" so that we could test if an issue was a hardware issue or a software one. Fortunately things have improved since then and the "Test Motors/Encoders" button under actions has replaced the need for a dedicated firmware. The firmware only works for the beta tester shelds anyway so it can't test any hardware except version 1.0 which is misleading. This PR deletes the file which could be confusing.