Fix for z not retracting before moving home
What's happening is that since checkForStopCommand() is not called unless the machine is moving a stop sent when the machine is not running causes issues.

The stop command (!) will sit in the buffer until a move command is sent. As soon as the move command is processed, the stop is detected and the line is not executed.

The solution is to also check for the stop command when the machine is not moving.