Remove Whole Number of Encoder Steps; Fix Default Values
I don't know how important this the whole number of encoder steps
really is.  It would be more of an issue if we used the KI term in
the PID calculations.  We can always add back later if we think it
actually contributes something important.

Also changed the default values of a few related variables, the
defaults are never used, but it should help future programers if
the defaults at least make some sense.