Create Realtime Loop; Perform System Init on Stop
Created a Realtime function that is called frequently to act on
realtime actions.  Put all realtime functions in here. This includes
returning position and error, reading the serial buffer.  Remove all
calls to these functions in the code.

Remove all calls to Delay in favor of maslowDelay which calls the
realtime function.

On Stop, return from all functions back to base loop and do a reset
similar to init.

Clean up some usage of passed variables where not necessary in
returnpoz

In RingBuffer use the passed buffer and don't constantly create
large 128 byte strings.

Create a number of precompile defines to make changing some things
easier and to remove need for variables in some places.