Remove CheckForStopCommand; Adjust HoldPosition
Stop command function is no longer needed.

HoldPosition still seems like an oddity, need to think about
what we really expect this to do.  Right now it just turns off
a motor if that motor hasn't received instructions in two seconds.
Perhaps, it should consider whether the motor is moving or has moved.
Do we care if we are in a motion sequence or not?  Perhaps eventually
it can be handled by the planner.