Implement Settings; Create ReadFloat; Version Track Steps
Add output listing all available settings.  Add logic for
processing a write of a setting.

Settings in total take about 165 bytes of space even with being a
bit lazy.  I thik we have 4096 to work with so we should be fine.

Disable help printout for items not enabled in maslow.

Add version tracking to step eeprom write.  This should enable
us to upgrade or alter the step eeprom structure and still
preserve the machine position.

Global settings will not be preserved between an update of the
settings struct however.  There are no associative arrays in
c++ and manually mapping old values to new doesn't seem worth
it at least not now.

Add readFloat function to replace findEndOfNumber.  The old
function used isdigit and is Punct to parse the number, however
ispunct allows for all sorts of punctuation beyond '+' '_' and '.'.
Including comment characters.  isDigit allows for A-E characters
which eat up a gcode word.