Integrate Settings; Add zAxisAttached Setting
Replaced updateKinematics and updateMotor functions since all of that
is handled by settings now.  Integrated the various settings into the
code.

Will have to figure out how all of the geometry recomputes and stuff
work in this design.

I also discovered that I had failed to add the zAxisAttached setting.
It rightly belonged near other zaxis settings which required a renumbering
of everything.  I wouldn't recommend this as it is reeeeaaaal tedious
to do.