Save Steps to EEPROM; Add Wipe EEPROM; Remove Junk
Removed the wipe eeprom from the axes and related B command.

Added $RST=[$,#,*] command.  $ wipes only settings but keeps
the step data, # wipes all maslow data, and * wipes the entire
eeprom.

Added a $K command.  Whenever changing a setting marked with $K
you will want to call this command.  This will cause the kinematics
to be recomputed.  It is not done automatically because that would
take a lot of time if GC was chaning each setting individually
and the user might receive a bunch of chain calibration messages.

Removed is machine ready and safe command concepts.  Since all
settings, at least default ones, are now on the machine, the machine
is always somewhat ready.  This saved a bunch of memory.