Add Support for Importing Prior Position Data
Prior position data is stored as a calculated value.
  Steps * encoderSteps / distancePerRotation

As a result, we can't load the prior values until we have
encoderSteps and distancePerRotation data from the user and
we can't rely on default data.

This now checks for old position data, and if it is found
sets a flag locking everything until the proper settings
are sent to the firmware.  At that point the data is
converted to the proper steps, saved, and the old data is
cleared from eeprom to prevent future issues.

This will only run if valid prior position data is found,
not if no data is found.

In order to enable all of this to work, moved the start of
current settings scheme to addres 300+.  Essentially wastes
299 bytes of EEPROM.  Annoying, but not the worst thing.