Add a Simple Fake Servo Option
This is very basic, it adds up to 10% error to mimic
servo noise.  It doesn't do a great job of testing the
PID settings (this would be tough) and doesn't verify
that the encoder library is working.  But other than
that I think this tests most of the code in the Firmware.