Fix for issue #379 (won't start again after stop)
Fixes issue #379, the machine won't start again after stop. Basically, the line is being cleared from the buffer when the stop command is recieved, but no OK is being sent to indicate the space is free. It seems counter intuitive to send an OK message when the machine is stopped, but the buffer space is available.

Is this the right fix?