Add support for running in simavr
Simavr is a simulator for the AVR family of microcontrollers
(https://github.com/buserror/simavr). Running Maslow firmware in simavr
is useful for automated testing in a CI loop and debugging in GDB.

These commit works around two issues with simavr:
1. The simulated UART crashes if you don't flush often, so we flush
after every character written to Serial.
2. Our TimerOne class causes the simulator to hang, so I moved the call
to runsOnATimer() into loop(). This means runsOnATimer() will be called
less frequently, but I think that's OK for the simulated environment.