rename 'zAxisAuto'
PR #328 added a setting to enable/disable spindle control. The setting was named 'zAxisAuto' which seems confusing as it is being used as an on/off control for spindle power. Change the name and accompanying text in Report.cpp to 'spindleAutomation' to better reflect its use. Note that the change in GC that accomanied #328 was PR #458.