Add a PlatformIO project
This project builds for two environments:
megaatmega2560: This is the standard Arduino Mega 2560 setup for
PlatformIO. It uses the default flags for PlatformIO, and builds an
optimized, stripped binary. Good for building firmware for production
use.
simavr: Disable all optimizations, leave debugging symbols, and #define
SIMAVR. SIMAVR doesn't do anything yet, but it will enable some
simavr-specific behavior in a separate commit.

I created this project as prep for #365, but it can also help simplify
setup for users of the PlatformIO IDE, or someone setting up a CI
environment.

I think it also resolves #365.