WIP: Disable uploading for simavr
NOTE: This is not fully tested, and is being uploaded for @blurfl to
take a look.

My previous commit created a PlatformIO project that built a customised
binary for running in simavr. It also attempted to upload this binary
when an Arduino was plugged in, which was never intended.

This commit disables the uploading for simavr.

Other environments