Add a system setting for positionErrorLimit, stop cutting if reached
 - add $42 positionErrorLimit with a default of 2mm
 - check each time [PE: is sent
 - set sys.stop flag if left- or right-axis error >= positionErrorLimit

Note, this leaves GC with a 'Hold' in place rather than a 'Stop'. This leaves the line number visible.The user is asked to click the 'Stop' button to get GC into the same state as the firmware.

Might want a GC PR to make an Advanced Setting available to change the value of positionErrorLimit.