Replace G2/G3 line with equivalent G1 line
The problem here is floating point precision, but it can be identified by looking at the height of the chord described by the G2/G3 command. Problematic commands have a chord that resolves to 0.01mm or less. This code trapps that situation and calls G1 instead using the endpoint. This turns an arc with less than 0.01mm of bow into a straight line.