Adds a link to a wiki page for when the sled is not keeping up
Since this error catches most issues we've been getting a lot of folks asking about it in the forums.

The PR adds the URL of a wiki page with more information to the popup. I also removed the words about writing down the line number because the popup occurs both when a program is running and when there is no program running.

I would really like the link to be clickable, but opening a web browser from a popup is prety tricky so this will have to work for now