B11 should not trigger position error alarm
The B11 gcode moves a motor at speed S for T seconds. It should not trigger the position error alarm as it is a testing gcode.

Usage example (moves the left motor at speed 200 for 2 seconds:
B11 L1 S200 T2

(S)peed accepts a float but treats that as an integer 0..255
(T)ime is a float