Change originalChainLength to 1651 mm
Though it really doesn't matter to do this here, because ground control overwrites the default values, this proposal changes the originalChainLength default value to 1651 mm, which is an integral of 6.35 mm (the pitch of a #25 chain) and therefore would result in a tooth being at 12 o'clock after the chains are extended.  Using this value can allow someone to manually reset the chains without having to go through the long, tedious process of having the motor feed out the chain.