Remove artificial last else
Hopefully better mod:
The artificial last `else` is removed with all its assignments.
The last check `if(pcbVersion == 3)`  is also removed so that the resulting else makes the compiler happy.
Unconnected pins are set to  -1 in the board versions <= 2 .
Version test in the dummy actions has been changed from `== 3` to `>= 3` to reflect the aforementioned change.