Initialize sys.feedrate to 400 mm/min
sys.feedrate was not being initialized and resulted in a feed rate of 1 mm/min when gcode was sent without a feed rate (if no other feed rate had been previously sent).  This sets the feedrate to 400 mm/min (half of the maximum feed rate).  It appears to me that the firmware works in metric so this should be appropriate .