Add Missing Kinematics.Init()
Any changes to values that are used during the kinematics::forward calculations should call kinematics.init() so sled position can be recalculated.  If its not, this can cause a "sled not keeping up" error, particularly after calibration.  If the sled is at an unadjusted position and a move is made, the move will be made with the adjusted parameters and that can exceed the positionErrorLimit.